Treat all headers as lower case
We force both key and value to lowercase. This is needed so we can
process for example Host header both as "Host:" and "host:", as well as
force-folding the value coming in since it's user specified.

Might need to make this configurable in the future, but for now the new
default seems ok.